requirements.txt
Django
django-tastypie
django_compressor
django-filter
